+++ lib/src/list_extensions.dart
@@ -1,5 +1,8 @@
 /// Extension methods for any [List].
 extension ListExtensions<E> on List<E> {
   /// Copy current list with adding [element] at the end of new list.
-  List<E> copyWith(E element) => List.from(this)..add(element);
+  ///
+  /// If current list is `null` - new list with [element] will be created.
+  List<E> copyWith(E element) =>
+      this == null ? [element] : (List.from(this)..add(element));
 }